ui-generator-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
python-multipart
Pillow